#45 Adjust new proxy for conventions
