+++ VNTextProxy/AlternateProxies/WinMM/exports.def
@@ -1,0 +1,182 @@
+LIBRARY winmm
+EXPORTS
+	CloseDriver=FakeCloseDriver @1
+	DefDriverProc=FakeDefDriverProc @2
+	DriverCallback=FakeDriverCallback @3
+	DrvGetModuleHandle=FakeDrvGetModuleHandle @4
+	GetDriverModuleHandle=FakeGetDriverModuleHandle @5
+	OpenDriver=FakeOpenDriver @6
+	PlaySound=FakePlaySound @7
+	PlaySoundA=FakePlaySoundA @8
+	PlaySoundW=FakePlaySoundW @9
+	SendDriverMessage=FakeSendDriverMessage @10
+	WOWAppExit=FakeWOWAppExit @11
+	auxGetDevCapsA=FakeauxGetDevCapsA @12
+	auxGetDevCapsW=FakeauxGetDevCapsW @13
+	auxGetNumDevs=FakeauxGetNumDevs @14
+	auxGetVolume=FakeauxGetVolume @15
+	auxOutMessage=FakeauxOutMessage @16
+	auxSetVolume=FakeauxSetVolume @17
+	joyConfigChanged=FakejoyConfigChanged @18
+	joyGetDevCapsA=FakejoyGetDevCapsA @19
+	joyGetDevCapsW=FakejoyGetDevCapsW @20
+	joyGetNumDevs=FakejoyGetNumDevs @21
+	joyGetPos=FakejoyGetPos @22
+	joyGetPosEx=FakejoyGetPosEx @23
+	joyGetThreshold=FakejoyGetThreshold @24
+	joyReleaseCapture=FakejoyReleaseCapture @25
+	joySetCapture=FakejoySetCapture @26
+	joySetThreshold=FakejoySetThreshold @27
+	mciDriverNotify=FakemciDriverNotify @28
+	mciDriverYield=FakemciDriverYield @29
+	mciExecute=FakemciExecute @30
+	mciFreeCommandResource=FakemciFreeCommandResource @31
+	mciGetCreatorTask=FakemciGetCreatorTask @32
+	mciGetDeviceIDA=FakemciGetDeviceIDA @33
+	mciGetDeviceIDFromElementIDA=FakemciGetDeviceIDFromElementIDA @34
+	mciGetDeviceIDFromElementIDW=FakemciGetDeviceIDFromElementIDW @35
+	mciGetDeviceIDW=FakemciGetDeviceIDW @36
+	mciGetDriverData=FakemciGetDriverData @37
+	mciGetErrorStringA=FakemciGetErrorStringA @38
+	mciGetErrorStringW=FakemciGetErrorStringW @39
+	mciGetYieldProc=FakemciGetYieldProc @40
+	mciLoadCommandResource=FakemciLoadCommandResource @41
+	mciSendCommandA=FakemciSendCommandA @42
+	mciSendCommandW=FakemciSendCommandW @43
+	mciSendStringA=FakemciSendStringA @44
+	mciSendStringW=FakemciSendStringW @45
+	mciSetDriverData=FakemciSetDriverData @46
+	mciSetYieldProc=FakemciSetYieldProc @47
+	midiConnect=FakemidiConnect @48
+	midiDisconnect=FakemidiDisconnect @49
+	midiInAddBuffer=FakemidiInAddBuffer @50
+	midiInClose=FakemidiInClose @51
+	midiInGetDevCapsA=FakemidiInGetDevCapsA @52
+	midiInGetDevCapsW=FakemidiInGetDevCapsW @53
+	midiInGetErrorTextA=FakemidiInGetErrorTextA @54
+	midiInGetErrorTextW=FakemidiInGetErrorTextW @55
+	midiInGetID=FakemidiInGetID @56
+	midiInGetNumDevs=FakemidiInGetNumDevs @57
+	midiInMessage=FakemidiInMessage @58
+	midiInOpen=FakemidiInOpen @59
+	midiInPrepareHeader=FakemidiInPrepareHeader @60
+	midiInReset=FakemidiInReset @61
+	midiInStart=FakemidiInStart @62
+	midiInStop=FakemidiInStop @63
+	midiInUnprepareHeader=FakemidiInUnprepareHeader @64
+	midiOutCacheDrumPatches=FakemidiOutCacheDrumPatches @65
+	midiOutCachePatches=FakemidiOutCachePatches @66
+	midiOutClose=FakemidiOutClose @67
+	midiOutGetDevCapsA=FakemidiOutGetDevCapsA @68
+	midiOutGetDevCapsW=FakemidiOutGetDevCapsW @69
+	midiOutGetErrorTextA=FakemidiOutGetErrorTextA @70
+	midiOutGetErrorTextW=FakemidiOutGetErrorTextW @71
+	midiOutGetID=FakemidiOutGetID @72
+	midiOutGetNumDevs=FakemidiOutGetNumDevs @73
+	midiOutGetVolume=FakemidiOutGetVolume @74
+	midiOutLongMsg=FakemidiOutLongMsg @75
+	midiOutMessage=FakemidiOutMessage @76
+	midiOutOpen=FakemidiOutOpen @77
+	midiOutPrepareHeader=FakemidiOutPrepareHeader @78
+	midiOutReset=FakemidiOutReset @79
+	midiOutSetVolume=FakemidiOutSetVolume @80
+	midiOutShortMsg=FakemidiOutShortMsg @81
+	midiOutUnprepareHeader=FakemidiOutUnprepareHeader @82
+	midiStreamClose=FakemidiStreamClose @83
+	midiStreamOpen=FakemidiStreamOpen @84
+	midiStreamOut=FakemidiStreamOut @85
+	midiStreamPause=FakemidiStreamPause @86
+	midiStreamPosition=FakemidiStreamPosition @87
+	midiStreamProperty=FakemidiStreamProperty @88
+	midiStreamRestart=FakemidiStreamRestart @89
+	midiStreamStop=FakemidiStreamStop @90
+	mixerClose=FakemixerClose @91
+	mixerGetControlDetailsA=FakemixerGetControlDetailsA @92
+	mixerGetControlDetailsW=FakemixerGetControlDetailsW @93
+	mixerGetDevCapsA=FakemixerGetDevCapsA @94
+	mixerGetDevCapsW=FakemixerGetDevCapsW @95
+	mixerGetID=FakemixerGetID @96
+	mixerGetLineControlsA=FakemixerGetLineControlsA @97
+	mixerGetLineControlsW=FakemixerGetLineControlsW @98
+	mixerGetLineInfoA=FakemixerGetLineInfoA @99
+	mixerGetLineInfoW=FakemixerGetLineInfoW @100
+	mixerGetNumDevs=FakemixerGetNumDevs @101
+	mixerMessage=FakemixerMessage @102
+	mixerOpen=FakemixerOpen @103
+	mixerSetControlDetails=FakemixerSetControlDetails @104
+	mmDrvInstall=FakemmDrvInstall @105
+	mmGetCurrentTask=FakemmGetCurrentTask @106
+	mmTaskBlock=FakemmTaskBlock @107
+	mmTaskCreate=FakemmTaskCreate @108
+	mmTaskSignal=FakemmTaskSignal @109
+	mmTaskYield=FakemmTaskYield @110
+	mmioAdvance=FakemmioAdvance @111
+	mmioAscend=FakemmioAscend @112
+	mmioClose=FakemmioClose @113
+	mmioCreateChunk=FakemmioCreateChunk @114
+	mmioDescend=FakemmioDescend @115
+	mmioFlush=FakemmioFlush @116
+	mmioGetInfo=FakemmioGetInfo @117
+	mmioInstallIOProcA=FakemmioInstallIOProcA @118
+	mmioInstallIOProcW=FakemmioInstallIOProcW @119
+	mmioOpenA=FakemmioOpenA @120
+	mmioOpenW=FakemmioOpenW @121
+	mmioRead=FakemmioRead @122
+	mmioRenameA=FakemmioRenameA @123
+	mmioRenameW=FakemmioRenameW @124
+	mmioSeek=FakemmioSeek @125
+	mmioSendMessage=FakemmioSendMessage @126
+	mmioSetBuffer=FakemmioSetBuffer @127
+	mmioSetInfo=FakemmioSetInfo @128
+	mmioStringToFOURCCA=FakemmioStringToFOURCCA @129
+	mmioStringToFOURCCW=FakemmioStringToFOURCCW @130
+	mmioWrite=FakemmioWrite @131
+	mmsystemGetVersion=FakemmsystemGetVersion @132
+	sndPlaySoundA=FakesndPlaySoundA @133
+	sndPlaySoundW=FakesndPlaySoundW @134
+	timeBeginPeriod=FaketimeBeginPeriod @135
+	timeEndPeriod=FaketimeEndPeriod @136
+	timeGetDevCaps=FaketimeGetDevCaps @137
+	timeGetSystemTime=FaketimeGetSystemTime @138
+	timeGetTime=FaketimeGetTime @139
+	timeKillEvent=FaketimeKillEvent @140
+	timeSetEvent=FaketimeSetEvent @141
+	waveInAddBuffer=FakewaveInAddBuffer @142
+	waveInClose=FakewaveInClose @143
+	waveInGetDevCapsA=FakewaveInGetDevCapsA @144
+	waveInGetDevCapsW=FakewaveInGetDevCapsW @145
+	waveInGetErrorTextA=FakewaveInGetErrorTextA @146
+	waveInGetErrorTextW=FakewaveInGetErrorTextW @147
+	waveInGetID=FakewaveInGetID @148
+	waveInGetNumDevs=FakewaveInGetNumDevs @149
+	waveInGetPosition=FakewaveInGetPosition @150
+	waveInMessage=FakewaveInMessage @151
+	waveInOpen=FakewaveInOpen @152
+	waveInPrepareHeader=FakewaveInPrepareHeader @153
+	waveInReset=FakewaveInReset @154
+	waveInStart=FakewaveInStart @155
+	waveInStop=FakewaveInStop @156
+	waveInUnprepareHeader=FakewaveInUnprepareHeader @157
+	waveOutBreakLoop=FakewaveOutBreakLoop @158
+	waveOutClose=FakewaveOutClose @159
+	waveOutGetDevCapsA=FakewaveOutGetDevCapsA @160
+	waveOutGetDevCapsW=FakewaveOutGetDevCapsW @161
+	waveOutGetErrorTextA=FakewaveOutGetErrorTextA @162
+	waveOutGetErrorTextW=FakewaveOutGetErrorTextW @163
+	waveOutGetID=FakewaveOutGetID @164
+	waveOutGetNumDevs=FakewaveOutGetNumDevs @165
+	waveOutGetPitch=FakewaveOutGetPitch @166
+	waveOutGetPlaybackRate=FakewaveOutGetPlaybackRate @167
+	waveOutGetPosition=FakewaveOutGetPosition @168
+	waveOutGetVolume=FakewaveOutGetVolume @169
+	waveOutMessage=FakewaveOutMessage @170
+	waveOutOpen=FakewaveOutOpen @171
+	waveOutPause=FakewaveOutPause @172
+	waveOutPrepareHeader=FakewaveOutPrepareHeader @173
+	waveOutReset=FakewaveOutReset @174
+	waveOutRestart=FakewaveOutRestart @175
+	waveOutSetPitch=FakewaveOutSetPitch @176
+	waveOutSetPlaybackRate=FakewaveOutSetPlaybackRate @177
+	waveOutSetVolume=FakewaveOutSetVolume @178
+	waveOutUnprepareHeader=FakewaveOutUnprepareHeader @179
+	waveOutWrite=FakewaveOutWrite @180
