--- VNTextProxy/AlternateProxies/winmm/Proxy.h
@@ -1,188 +1,0 @@
-#pragma once
-
-class Proxy
-{
-public:
-    static void Init();
-	
-	static inline void* OriginalCloseDriver{};
-	static inline void* OriginalDefDriverProc{};
-	static inline void* OriginalDriverCallback{};
-	static inline void* OriginalDrvGetModuleHandle{};
-	static inline void* OriginalGetDriverModuleHandle{};
-	static inline void* OriginalOpenDriver{};
-	static inline void* OriginalPlaySound{};
-	static inline void* OriginalPlaySoundA{};
-	static inline void* OriginalPlaySoundW{};
-	static inline void* OriginalSendDriverMessage{};
-	static inline void* OriginalWOWAppExit{};
-	static inline void* OriginalauxGetDevCapsA{};
-	static inline void* OriginalauxGetDevCapsW{};
-	static inline void* OriginalauxGetNumDevs{};
-	static inline void* OriginalauxGetVolume{};
-	static inline void* OriginalauxOutMessage{};
-	static inline void* OriginalauxSetVolume{};
-	static inline void* OriginaljoyConfigChanged{};
-	static inline void* OriginaljoyGetDevCapsA{};
-	static inline void* OriginaljoyGetDevCapsW{};
-	static inline void* OriginaljoyGetNumDevs{};
-	static inline void* OriginaljoyGetPos{};
-	static inline void* OriginaljoyGetPosEx{};
-	static inline void* OriginaljoyGetThreshold{};
-	static inline void* OriginaljoyReleaseCapture{};
-	static inline void* OriginaljoySetCapture{};
-	static inline void* OriginaljoySetThreshold{};
-	static inline void* OriginalmciDriverNotify{};
-	static inline void* OriginalmciDriverYield{};
-	static inline void* OriginalmciExecute{};
-	static inline void* OriginalmciFreeCommandResource{};
-	static inline void* OriginalmciGetCreatorTask{};
-	static inline void* OriginalmciGetDeviceIDA{};
-	static inline void* OriginalmciGetDeviceIDFromElementIDA{};
-	static inline void* OriginalmciGetDeviceIDFromElementIDW{};
-	static inline void* OriginalmciGetDeviceIDW{};
-	static inline void* OriginalmciGetDriverData{};
-	static inline void* OriginalmciGetErrorStringA{};
-	static inline void* OriginalmciGetErrorStringW{};
-	static inline void* OriginalmciGetYieldProc{};
-	static inline void* OriginalmciLoadCommandResource{};
-	static inline void* OriginalmciSendCommandA{};
-	static inline void* OriginalmciSendCommandW{};
-	static inline void* OriginalmciSendStringA{};
-	static inline void* OriginalmciSendStringW{};
-	static inline void* OriginalmciSetDriverData{};
-	static inline void* OriginalmciSetYieldProc{};
-	static inline void* OriginalmidiConnect{};
-	static inline void* OriginalmidiDisconnect{};
-	static inline void* OriginalmidiInAddBuffer{};
-	static inline void* OriginalmidiInClose{};
-	static inline void* OriginalmidiInGetDevCapsA{};
-	static inline void* OriginalmidiInGetDevCapsW{};
-	static inline void* OriginalmidiInGetErrorTextA{};
-	static inline void* OriginalmidiInGetErrorTextW{};
-	static inline void* OriginalmidiInGetID{};
-	static inline void* OriginalmidiInGetNumDevs{};
-	static inline void* OriginalmidiInMessage{};
-	static inline void* OriginalmidiInOpen{};
-	static inline void* OriginalmidiInPrepareHeader{};
-	static inline void* OriginalmidiInReset{};
-	static inline void* OriginalmidiInStart{};
-	static inline void* OriginalmidiInStop{};
-	static inline void* OriginalmidiInUnprepareHeader{};
-	static inline void* OriginalmidiOutCacheDrumPatches{};
-	static inline void* OriginalmidiOutCachePatches{};
-	static inline void* OriginalmidiOutClose{};
-	static inline void* OriginalmidiOutGetDevCapsA{};
-	static inline void* OriginalmidiOutGetDevCapsW{};
-	static inline void* OriginalmidiOutGetErrorTextA{};
-	static inline void* OriginalmidiOutGetErrorTextW{};
-	static inline void* OriginalmidiOutGetID{};
-	static inline void* OriginalmidiOutGetNumDevs{};
-	static inline void* OriginalmidiOutGetVolume{};
-	static inline void* OriginalmidiOutLongMsg{};
-	static inline void* OriginalmidiOutMessage{};
-	static inline void* OriginalmidiOutOpen{};
-	static inline void* OriginalmidiOutPrepareHeader{};
-	static inline void* OriginalmidiOutReset{};
-	static inline void* OriginalmidiOutSetVolume{};
-	static inline void* OriginalmidiOutShortMsg{};
-	static inline void* OriginalmidiOutUnprepareHeader{};
-	static inline void* OriginalmidiStreamClose{};
-	static inline void* OriginalmidiStreamOpen{};
-	static inline void* OriginalmidiStreamOut{};
-	static inline void* OriginalmidiStreamPause{};
-	static inline void* OriginalmidiStreamPosition{};
-	static inline void* OriginalmidiStreamProperty{};
-	static inline void* OriginalmidiStreamRestart{};
-	static inline void* OriginalmidiStreamStop{};
-	static inline void* OriginalmixerClose{};
-	static inline void* OriginalmixerGetControlDetailsA{};
-	static inline void* OriginalmixerGetControlDetailsW{};
-	static inline void* OriginalmixerGetDevCapsA{};
-	static inline void* OriginalmixerGetDevCapsW{};
-	static inline void* OriginalmixerGetID{};
-	static inline void* OriginalmixerGetLineControlsA{};
-	static inline void* OriginalmixerGetLineControlsW{};
-	static inline void* OriginalmixerGetLineInfoA{};
-	static inline void* OriginalmixerGetLineInfoW{};
-	static inline void* OriginalmixerGetNumDevs{};
-	static inline void* OriginalmixerMessage{};
-	static inline void* OriginalmixerOpen{};
-	static inline void* OriginalmixerSetControlDetails{};
-	static inline void* OriginalmmDrvInstall{};
-	static inline void* OriginalmmGetCurrentTask{};
-	static inline void* OriginalmmTaskBlock{};
-	static inline void* OriginalmmTaskCreate{};
-	static inline void* OriginalmmTaskSignal{};
-	static inline void* OriginalmmTaskYield{};
-	static inline void* OriginalmmioAdvance{};
-	static inline void* OriginalmmioAscend{};
-	static inline void* OriginalmmioClose{};
-	static inline void* OriginalmmioCreateChunk{};
-	static inline void* OriginalmmioDescend{};
-	static inline void* OriginalmmioFlush{};
-	static inline void* OriginalmmioGetInfo{};
-	static inline void* OriginalmmioInstallIOProcA{};
-	static inline void* OriginalmmioInstallIOProcW{};
-	static inline void* OriginalmmioOpenA{};
-	static inline void* OriginalmmioOpenW{};
-	static inline void* OriginalmmioRead{};
-	static inline void* OriginalmmioRenameA{};
-	static inline void* OriginalmmioRenameW{};
-	static inline void* OriginalmmioSeek{};
-	static inline void* OriginalmmioSendMessage{};
-	static inline void* OriginalmmioSetBuffer{};
-	static inline void* OriginalmmioSetInfo{};
-	static inline void* OriginalmmioStringToFOURCCA{};
-	static inline void* OriginalmmioStringToFOURCCW{};
-	static inline void* OriginalmmioWrite{};
-	static inline void* OriginalmmsystemGetVersion{};
-	static inline void* OriginalsndPlaySoundA{};
-	static inline void* OriginalsndPlaySoundW{};
-	static inline void* OriginaltimeBeginPeriod{};
-	static inline void* OriginaltimeEndPeriod{};
-	static inline void* OriginaltimeGetDevCaps{};
-	static inline void* OriginaltimeGetSystemTime{};
-	static inline void* OriginaltimeGetTime{};
-	static inline void* OriginaltimeKillEvent{};
-	static inline void* OriginaltimeSetEvent{};
-	static inline void* OriginalwaveInAddBuffer{};
-	static inline void* OriginalwaveInClose{};
-	static inline void* OriginalwaveInGetDevCapsA{};
-	static inline void* OriginalwaveInGetDevCapsW{};
-	static inline void* OriginalwaveInGetErrorTextA{};
-	static inline void* OriginalwaveInGetErrorTextW{};
-	static inline void* OriginalwaveInGetID{};
-	static inline void* OriginalwaveInGetNumDevs{};
-	static inline void* OriginalwaveInGetPosition{};
-	static inline void* OriginalwaveInMessage{};
-	static inline void* OriginalwaveInOpen{};
-	static inline void* OriginalwaveInPrepareHeader{};
-	static inline void* OriginalwaveInReset{};
-	static inline void* OriginalwaveInStart{};
-	static inline void* OriginalwaveInStop{};
-	static inline void* OriginalwaveInUnprepareHeader{};
-	static inline void* OriginalwaveOutBreakLoop{};
-	static inline void* OriginalwaveOutClose{};
-	static inline void* OriginalwaveOutGetDevCapsA{};
-	static inline void* OriginalwaveOutGetDevCapsW{};
-	static inline void* OriginalwaveOutGetErrorTextA{};
-	static inline void* OriginalwaveOutGetErrorTextW{};
-	static inline void* OriginalwaveOutGetID{};
-	static inline void* OriginalwaveOutGetNumDevs{};
-	static inline void* OriginalwaveOutGetPitch{};
-	static inline void* OriginalwaveOutGetPlaybackRate{};
-	static inline void* OriginalwaveOutGetPosition{};
-	static inline void* OriginalwaveOutGetVolume{};
-	static inline void* OriginalwaveOutMessage{};
-	static inline void* OriginalwaveOutOpen{};
-	static inline void* OriginalwaveOutPause{};
-	static inline void* OriginalwaveOutPrepareHeader{};
-	static inline void* OriginalwaveOutReset{};
-	static inline void* OriginalwaveOutRestart{};
-	static inline void* OriginalwaveOutSetPitch{};
-	static inline void* OriginalwaveOutSetPlaybackRate{};
-	static inline void* OriginalwaveOutSetVolume{};
-	static inline void* OriginalwaveOutUnprepareHeader{};
-	static inline void* OriginalwaveOutWrite{};
-};
