--- VNTextProxy/AlternateProxies/winmm/exports.def
@@ -1,182 +1,0 @@
-LIBRARY winmm
-EXPORTS
-	CloseDriver=fCloseDriver @1
-	DefDriverProc=fDefDriverProc @2
-	DriverCallback=fDriverCallback @3
-	DrvGetModuleHandle=fDrvGetModuleHandle @4
-	GetDriverModuleHandle=fGetDriverModuleHandle @5
-	OpenDriver=fOpenDriver @6
-	PlaySound=fPlaySound @7
-	PlaySoundA=fPlaySoundA @8
-	PlaySoundW=fPlaySoundW @9
-	SendDriverMessage=fSendDriverMessage @10
-	WOWAppExit=fWOWAppExit @11
-	auxGetDevCapsA=fauxGetDevCapsA @12
-	auxGetDevCapsW=fauxGetDevCapsW @13
-	auxGetNumDevs=fauxGetNumDevs @14
-	auxGetVolume=fauxGetVolume @15
-	auxOutMessage=fauxOutMessage @16
-	auxSetVolume=fauxSetVolume @17
-	joyConfigChanged=fjoyConfigChanged @18
-	joyGetDevCapsA=fjoyGetDevCapsA @19
-	joyGetDevCapsW=fjoyGetDevCapsW @20
-	joyGetNumDevs=fjoyGetNumDevs @21
-	joyGetPos=fjoyGetPos @22
-	joyGetPosEx=fjoyGetPosEx @23
-	joyGetThreshold=fjoyGetThreshold @24
-	joyReleaseCapture=fjoyReleaseCapture @25
-	joySetCapture=fjoySetCapture @26
-	joySetThreshold=fjoySetThreshold @27
-	mciDriverNotify=fmciDriverNotify @28
-	mciDriverYield=fmciDriverYield @29
-	mciExecute=fmciExecute @30
-	mciFreeCommandResource=fmciFreeCommandResource @31
-	mciGetCreatorTask=fmciGetCreatorTask @32
-	mciGetDeviceIDA=fmciGetDeviceIDA @33
-	mciGetDeviceIDFromElementIDA=fmciGetDeviceIDFromElementIDA @34
-	mciGetDeviceIDFromElementIDW=fmciGetDeviceIDFromElementIDW @35
-	mciGetDeviceIDW=fmciGetDeviceIDW @36
-	mciGetDriverData=fmciGetDriverData @37
-	mciGetErrorStringA=fmciGetErrorStringA @38
-	mciGetErrorStringW=fmciGetErrorStringW @39
-	mciGetYieldProc=fmciGetYieldProc @40
-	mciLoadCommandResource=fmciLoadCommandResource @41
-	mciSendCommandA=fmciSendCommandA @42
-	mciSendCommandW=fmciSendCommandW @43
-	mciSendStringA=fmciSendStringA @44
-	mciSendStringW=fmciSendStringW @45
-	mciSetDriverData=fmciSetDriverData @46
-	mciSetYieldProc=fmciSetYieldProc @47
-	midiConnect=fmidiConnect @48
-	midiDisconnect=fmidiDisconnect @49
-	midiInAddBuffer=fmidiInAddBuffer @50
-	midiInClose=fmidiInClose @51
-	midiInGetDevCapsA=fmidiInGetDevCapsA @52
-	midiInGetDevCapsW=fmidiInGetDevCapsW @53
-	midiInGetErrorTextA=fmidiInGetErrorTextA @54
-	midiInGetErrorTextW=fmidiInGetErrorTextW @55
-	midiInGetID=fmidiInGetID @56
-	midiInGetNumDevs=fmidiInGetNumDevs @57
-	midiInMessage=fmidiInMessage @58
-	midiInOpen=fmidiInOpen @59
-	midiInPrepareHeader=fmidiInPrepareHeader @60
-	midiInReset=fmidiInReset @61
-	midiInStart=fmidiInStart @62
-	midiInStop=fmidiInStop @63
-	midiInUnprepareHeader=fmidiInUnprepareHeader @64
-	midiOutCacheDrumPatches=fmidiOutCacheDrumPatches @65
-	midiOutCachePatches=fmidiOutCachePatches @66
-	midiOutClose=fmidiOutClose @67
-	midiOutGetDevCapsA=fmidiOutGetDevCapsA @68
-	midiOutGetDevCapsW=fmidiOutGetDevCapsW @69
-	midiOutGetErrorTextA=fmidiOutGetErrorTextA @70
-	midiOutGetErrorTextW=fmidiOutGetErrorTextW @71
-	midiOutGetID=fmidiOutGetID @72
-	midiOutGetNumDevs=fmidiOutGetNumDevs @73
-	midiOutGetVolume=fmidiOutGetVolume @74
-	midiOutLongMsg=fmidiOutLongMsg @75
-	midiOutMessage=fmidiOutMessage @76
-	midiOutOpen=fmidiOutOpen @77
-	midiOutPrepareHeader=fmidiOutPrepareHeader @78
-	midiOutReset=fmidiOutReset @79
-	midiOutSetVolume=fmidiOutSetVolume @80
-	midiOutShortMsg=fmidiOutShortMsg @81
-	midiOutUnprepareHeader=fmidiOutUnprepareHeader @82
-	midiStreamClose=fmidiStreamClose @83
-	midiStreamOpen=fmidiStreamOpen @84
-	midiStreamOut=fmidiStreamOut @85
-	midiStreamPause=fmidiStreamPause @86
-	midiStreamPosition=fmidiStreamPosition @87
-	midiStreamProperty=fmidiStreamProperty @88
-	midiStreamRestart=fmidiStreamRestart @89
-	midiStreamStop=fmidiStreamStop @90
-	mixerClose=fmixerClose @91
-	mixerGetControlDetailsA=fmixerGetControlDetailsA @92
-	mixerGetControlDetailsW=fmixerGetControlDetailsW @93
-	mixerGetDevCapsA=fmixerGetDevCapsA @94
-	mixerGetDevCapsW=fmixerGetDevCapsW @95
-	mixerGetID=fmixerGetID @96
-	mixerGetLineControlsA=fmixerGetLineControlsA @97
-	mixerGetLineControlsW=fmixerGetLineControlsW @98
-	mixerGetLineInfoA=fmixerGetLineInfoA @99
-	mixerGetLineInfoW=fmixerGetLineInfoW @100
-	mixerGetNumDevs=fmixerGetNumDevs @101
-	mixerMessage=fmixerMessage @102
-	mixerOpen=fmixerOpen @103
-	mixerSetControlDetails=fmixerSetControlDetails @104
-	mmDrvInstall=fmmDrvInstall @105
-	mmGetCurrentTask=fmmGetCurrentTask @106
-	mmTaskBlock=fmmTaskBlock @107
-	mmTaskCreate=fmmTaskCreate @108
-	mmTaskSignal=fmmTaskSignal @109
-	mmTaskYield=fmmTaskYield @110
-	mmioAdvance=fmmioAdvance @111
-	mmioAscend=fmmioAscend @112
-	mmioClose=fmmioClose @113
-	mmioCreateChunk=fmmioCreateChunk @114
-	mmioDescend=fmmioDescend @115
-	mmioFlush=fmmioFlush @116
-	mmioGetInfo=fmmioGetInfo @117
-	mmioInstallIOProcA=fmmioInstallIOProcA @118
-	mmioInstallIOProcW=fmmioInstallIOProcW @119
-	mmioOpenA=fmmioOpenA @120
-	mmioOpenW=fmmioOpenW @121
-	mmioRead=fmmioRead @122
-	mmioRenameA=fmmioRenameA @123
-	mmioRenameW=fmmioRenameW @124
-	mmioSeek=fmmioSeek @125
-	mmioSendMessage=fmmioSendMessage @126
-	mmioSetBuffer=fmmioSetBuffer @127
-	mmioSetInfo=fmmioSetInfo @128
-	mmioStringToFOURCCA=fmmioStringToFOURCCA @129
-	mmioStringToFOURCCW=fmmioStringToFOURCCW @130
-	mmioWrite=fmmioWrite @131
-	mmsystemGetVersion=fmmsystemGetVersion @132
-	sndPlaySoundA=fsndPlaySoundA @133
-	sndPlaySoundW=fsndPlaySoundW @134
-	timeBeginPeriod=ftimeBeginPeriod @135
-	timeEndPeriod=ftimeEndPeriod @136
-	timeGetDevCaps=ftimeGetDevCaps @137
-	timeGetSystemTime=ftimeGetSystemTime @138
-	timeGetTime=ftimeGetTime @139
-	timeKillEvent=ftimeKillEvent @140
-	timeSetEvent=ftimeSetEvent @141
-	waveInAddBuffer=fwaveInAddBuffer @142
-	waveInClose=fwaveInClose @143
-	waveInGetDevCapsA=fwaveInGetDevCapsA @144
-	waveInGetDevCapsW=fwaveInGetDevCapsW @145
-	waveInGetErrorTextA=fwaveInGetErrorTextA @146
-	waveInGetErrorTextW=fwaveInGetErrorTextW @147
-	waveInGetID=fwaveInGetID @148
-	waveInGetNumDevs=fwaveInGetNumDevs @149
-	waveInGetPosition=fwaveInGetPosition @150
-	waveInMessage=fwaveInMessage @151
-	waveInOpen=fwaveInOpen @152
-	waveInPrepareHeader=fwaveInPrepareHeader @153
-	waveInReset=fwaveInReset @154
-	waveInStart=fwaveInStart @155
-	waveInStop=fwaveInStop @156
-	waveInUnprepareHeader=fwaveInUnprepareHeader @157
-	waveOutBreakLoop=fwaveOutBreakLoop @158
-	waveOutClose=fwaveOutClose @159
-	waveOutGetDevCapsA=fwaveOutGetDevCapsA @160
-	waveOutGetDevCapsW=fwaveOutGetDevCapsW @161
-	waveOutGetErrorTextA=fwaveOutGetErrorTextA @162
-	waveOutGetErrorTextW=fwaveOutGetErrorTextW @163
-	waveOutGetID=fwaveOutGetID @164
-	waveOutGetNumDevs=fwaveOutGetNumDevs @165
-	waveOutGetPitch=fwaveOutGetPitch @166
-	waveOutGetPlaybackRate=fwaveOutGetPlaybackRate @167
-	waveOutGetPosition=fwaveOutGetPosition @168
-	waveOutGetVolume=fwaveOutGetVolume @169
-	waveOutMessage=fwaveOutMessage @170
-	waveOutOpen=fwaveOutOpen @171
-	waveOutPause=fwaveOutPause @172
-	waveOutPrepareHeader=fwaveOutPrepareHeader @173
-	waveOutReset=fwaveOutReset @174
-	waveOutRestart=fwaveOutRestart @175
-	waveOutSetPitch=fwaveOutSetPitch @176
-	waveOutSetPlaybackRate=fwaveOutSetPlaybackRate @177
-	waveOutSetVolume=fwaveOutSetVolume @178
-	waveOutUnprepareHeader=fwaveOutUnprepareHeader @179
-	waveOutWrite=fwaveOutWrite @180
